refactor: colocate hub prompt wrap helper and simplify tests
Move wrap_manifest_for_hub_push into _internal/_hub, drop redundant
create_commit mock tests, and add focused test_hub_manifest coverage.
