+++ python/langsmith/async_client.py
@@ -23,6 +23,7 @@ from typing import (
 import httpx
 
 from langsmith import client as ls_client
+from langsmith._internal._hub import wrap_manifest_for_hub_push
 from langsmith import schemas as ls_schemas
 from langsmith import utils as ls_utils
 from langsmith._internal import _profiles
@@ -1939,7 +1940,7 @@ class AsyncClient:
             json_object = dumps(prepped)
             manifest_dict = json.loads(json_object)
 
-        manifest_dict = ls_client.wrap_manifest_for_hub_push(manifest_dict)
+        manifest_dict = wrap_manifest_for_hub_push(manifest_dict)
 
         owner, prompt_name, _ = ls_utils.parse_prompt_identifier(prompt_identifier)
         prompt_owner_and_name = f"{owner}/{prompt_name}"
